Move voc and voi legend to bottom
* Use acronyms to save space

* Add variants label above strains axis

* Change sample size label to N

* Some style changes
